+++ extensions/menubar-weather/CHANGELOG.md
@@ -1,3 +1,10 @@
 # Menubar Weather Changelog
 
-## [Initial Version] - 2022-11-22
+## [Add forecast] - 2022-11-24
+
+- Add 7-day weather forecast
+- Add screencast
+
+## [Initial Version] - 2022-11-23
+
+- Show current weather on the menu bar